+++ NEWS.md
@@ -1,3 +1,5 @@
 # vctrs 0.1.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.
+
+* `vec_c()` and `vec_rbind()` now handle data.frame columns properly (@yutannihilation, #182)